+++ html/src/glsl/shaders/normal2color.glsl
@@ -1,0 +1,25 @@
+//#name normal2color
+//#vertex
+attribute vec3 aNormal;
+attribute vec3 aPosition;
+
+uniform mat4 uMatrix;
+uniform mat4 uViewMatrix;
+uniform mat4 uProjection;
+uniform mat3 uNormalMatrix;
+
+varying vec3 vNormal;
+	 
+void main() {
+	vNormal = uNormalMatrix * aNormal;
+	gl_Position = uProjection * uViewMatrix * vec4(aPosition, 1.0);
+}
+
+//#fragment
+precision mediump float;
+
+varying vec3 vNormal;
+
+void main() {
+	gl_FragColor = vec4(0.5 + 0.5 * vNormal, 1.0);
+}